gmscompat: Add baseline compatibility for unprivileged Play Services
Tested functionality:
  - Account login (including 2FA with NFC security key)
  - Play Store
  - Firebase Cloud Messaging (Signal, Discord, Slack, Gmail)
  - Firebase database API (Swift Backup)
  - Firebase app indexing (GMS debug settings)
  - Google Play Games
  - Account settings
  - Google My Account
  - Autofill
  - SMS verification receiver (Signal)
  - Play license verification (both in-app purchases and paid apps)
  - Play Store app purchases

Known issues:
  - Play feature delivery
  - SafetyNet doesn't pass

To open Play Services and account settings:
      am start com.google.android.gms/.app.settings.GoogleSettingsIALink

+++ core/java/android/app/NotificationManager.java
@@ -51,6 +51,8 @@ import android.service.notification.ZenPolicy;
 import android.util.Log;
 import android.util.proto.ProtoOutputStream;
 
+import com.android.internal.gmscompat.GmsHooks;
+
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.util.ArrayList;
@@ -901,6 +903,10 @@ public class NotificationManager {
      * had before it was deleted.
      */
     public void deleteNotificationChannel(String channelId) {
+        if (GmsHooks.skipDeleteNotificationChannel(channelId)) {
+            return;
+        }
+
         INotificationManager service = getService();
         try {
             service.deleteNotificationChannel(mContext.getPackageName(), channelId);
